+++ examples/file_generation_mode/lib/union/simple.dart
@@ -4,7 +4,7 @@ import 'package:file_generation_mode/pretty_print.dart';
 part 'simple.gen.dart';
 
 @Union()
-abstract class AsyncResult {
+sealed class AsyncResult {
   const AsyncResult._();
 
   /// Shorthand constructor