+++ package/lib/src/analyzer_plugin/analyzer_plugin_starter.dart
@@ -5,7 +5,7 @@ import 'package:analyzer_plugin/starter.dart';
 import 'package:data_class_plugin/src/analyzer_plugin/analyzer_plugin.dart';
 
 void start(Iterable<String> args, SendPort sendPort) {
-  ServerPluginStarter(DataClassPlugin(
+  ServerPluginStarter(DcpAnalyzerPlugin(
     PhysicalResourceProvider.INSTANCE,
   )).start(sendPort);
 }