+++ package/lib/src/annotations/union.dart
@@ -26,6 +26,8 @@ class Union {
     this.unionJsonKey,
     this.unionFallbackJsonValue,
     this.unmodifiableCollections,
+    this.toJsonNamedKeyToFactoryName,
+    this.when,
   });
 
   /// Toggles code generation for copyWith
@@ -64,6 +66,35 @@ class Union {
   final String? unionFallbackJsonValue;
 
   final bool? unmodifiableCollections;
+
+  /// Wraps the union fields into a new object with a key named after the factory
+  ///
+  /// ```
+  /// @Union()
+  /// class Result{
+  ///   factory Result.data(int value) = ResultData;
+  ///
+  ///   Map<String, dynamic> toJson();
+  /// }
+  /// ...
+  ///
+  /// // toJson call this produce
+  ///
+  /// {
+  ///   "data": {
+  ///     "value": value
+  ///   }
+  /// }
+  ///
+  /// instead of
+  ///
+  /// {
+  ///   "value": value
+  /// }
+  /// ```
+  final String? toJsonNamedKeyToFactoryName;
+
+  final bool? when;
 }
 
 class UnionJsonKeyValue {