+++ package/lib/src/backend/code_generator.dart
@@ -1,405 +1,99 @@
 import 'dart:async';
-import 'dart:io';
 
-import 'package:analyzer/dart/analysis/features.dart';
-import 'package:analyzer/dart/ast/ast.dart';
-import 'package:dart_style/dart_style.dart';
 import 'package:data_class_plugin/src/annotations/constants.dart';
-import 'package:data_class_plugin/src/backend/core/custom_dart_object.dart';
-import 'package:data_class_plugin/src/backend/core/declaration_finder.dart';
-import 'package:data_class_plugin/src/backend/core/declaration_info.dart';
-import 'package:data_class_plugin/src/backend/core/find_package_path_by_import.dart';
 import 'package:data_class_plugin/src/backend/core/generators/generators.dart';
 import 'package:data_class_plugin/src/backend/core/generators/union_from_json.dart';
 import 'package:data_class_plugin/src/backend/core/generators/union_when.dart';
-import 'package:data_class_plugin/src/backend/core/packages_dependency_graph.dart';
-import 'package:data_class_plugin/src/backend/core/parse_file_extension.dart';
-import 'package:data_class_plugin/src/backend/core/parsed_file_data.dart';
-import 'package:data_class_plugin/src/backend/core/parsed_files_registry.dart';
 import 'package:data_class_plugin/src/backend/core/utils.dart' as utils;
-import 'package:data_class_plugin/src/common/code_writer.dart';
-import 'package:data_class_plugin/src/common/utils.dart' as common_utils;
 import 'package:data_class_plugin/src/common/utils.dart';
 import 'package:data_class_plugin/src/extensions/extensions.dart';
 import 'package:data_class_plugin/src/options/data_class_plugin_options.dart';
-import 'package:data_class_plugin/src/tools/logger/plugin_logger.dart';
 import 'package:data_class_plugin/src/typedefs.dart';
-import 'package:data_class_plugin/src/visitors/visitors.dart';
-import 'package:glob/glob.dart';
 import 'package:path/path.dart' as path;
-import 'package:rxdart/rxdart.dart';
-import 'package:watcher/watcher.dart';
+import 'package:tachyon/tachyon.dart';
 
-final RegExp _dartFileNameMatcher = RegExp(r'^[a-zA-Z0-9_]+.dart$');
-
-class CodeGenerator {
-  CodeGenerator({
-    required this.directory,
-    PluginLogger? logger,
-  })  : _watcher = DirectoryWatcher(directory.path),
-        _logger = logger ?? PluginLogger();
-
-  final Directory directory;
-  final PluginLogger _logger;
-  final Watcher _watcher;
-
-  final Map<String, Completer<void>?> _activeWrites = <String, Completer<void>?>{};
-  final DependencyGraph _dependencyGraph = DependencyGraph();
-  final ParsedFilesRegistry _filesRegistry = ParsedFilesRegistry();
-
-  late final DeclarationFinder _declarationFinder = DeclarationFinder(
-    projectDirectoryPath: directory.path,
-    parsedFilesRegistry: _filesRegistry,
-    dependencyGraph: _dependencyGraph,
-  );
-
-  StreamSubscription<WatchEvent>? _watchSubscription;
-
-  List<String> get registeredFiles => _filesRegistry.keys.toList(growable: false);
-
-  /// Watches this project for any files changes and rebuilds when necessary
-  Future<void> watchProject({
-    void Function()? onReady,
-  }) async {
-    final Completer<void> completer = Completer<void>();
-    await indexProject();
-    await buildProject();
-    _watchSubscription = _watcher.events
-        .debounceTime(const Duration(milliseconds: 100))
-        .listen(_onWatchEvent, onDone: () {
-      completer.complete();
-    });
-    await _watcher.ready;
-    onReady?.call();
-    return completer.future;
-  }
-
-  Future<void> dispose() async {
-    await _watchSubscription?.cancel();
-  }
-
-  /// Indexes the project and creates links between source files
-  ///
-  /// [force] clears any previous indexing
-  Future<void> indexProject({bool forceClear = false}) async {
-    _logger.info('~ Indexing project..');
-
-    if (forceClear) {
-      _filesRegistry.clear();
-      _dependencyGraph.clear();
-    }
-
-    final Stopwatch stopwatch = Stopwatch()..start();
-
-    final DataClassPluginOptions pluginOptions =
-        await DataClassPluginOptions.fromFile(getDataClassPluginOptionsFile(directory.path));
-
-    final Iterable<File> dartFiles = directory
-        .listSync(recursive: true) //
-        .where((FileSystemEntity entity) {
-      if (entity is! File || !_dartFileNameMatcher.hasMatch(path.basename(entity.path))) {
-        return false;
-      }
-
-      return pluginOptions.allowedFilesGenerationPaths.any((Glob glob) {
-        return glob.matches(
-          path.relative(entity.absolute.path, from: directory.absolute.path),
-        );
-      });
-    }).cast<File>();
-
-    Future<void> index({
-      required String targetFilePath,
-      required CompilationUnit compilationUnit,
-    }) async {
-      for (final Directive directive in compilationUnit.directives) {
-        String? directiveUri;
-        if (directive is NamespaceDirective) {
-          directiveUri = directive.uri.stringValue;
-        }
-
-        if (directiveUri == null) {
-          continue;
-        }
-
-        final String? dartFilePath = await findDartFileFromUri(
-          projectDirectoryPath: directory.path,
-          currentDirectoryPath: File(targetFilePath).parent.absolute.path,
-          uri: directiveUri,
-        );
-
-        if (dartFilePath == null || !File(dartFilePath).existsSync()) {
-          continue;
-        }
-
-        if (_filesRegistry.containsKey(dartFilePath)) {
-          _dependencyGraph.add(
-            targetFilePath,
-            dartFilePath,
-          );
-          continue;
-        }
-
-        if (path.isWithin(directory.path, dartFilePath)) {
-          _dependencyGraph.add(targetFilePath, dartFilePath);
-          _filesRegistry[dartFilePath] = ParsedFileData(
-            absolutePath: dartFilePath,
-            compilationUnit:
-                dartFilePath.parse(featureSet: FeatureSet.latestLanguageVersion()).unit,
-            lastModifiedAt: File(dartFilePath).lastModifiedSync(),
-          );
-
-          await index(
-            targetFilePath: dartFilePath,
-            compilationUnit: _filesRegistry[dartFilePath]!.compilationUnit,
-          );
-        }
-      }
-    }
-
-    for (final File file in dartFiles) {
-      final String targetFilePath = file.absolute.path;
-
-      if (_filesRegistry.containsKey(targetFilePath)) {
-        continue;
-      }
-
-      _filesRegistry[targetFilePath] = ParsedFileData(
-        absolutePath: targetFilePath,
-        compilationUnit: targetFilePath.parse(featureSet: FeatureSet.latestLanguageVersion()).unit,
-        lastModifiedAt: file.lastModifiedSync(),
-      );
-
-      await index(
-        targetFilePath: targetFilePath,
-        compilationUnit: _filesRegistry[targetFilePath]!.compilationUnit,
-      );
-    }
-
-    stopwatch.stop();
-    _logger.info('~ Indexed ${_filesRegistry.length} files in ${stopwatch.elapsedMilliseconds}ms');
-  }
-
-  /// Builds the project
-  ///
-  /// **Project must be indexed before it can correctly build**
-  Future<void> buildProject() async {
-    _logger.info('~ Building project..');
-    final Stopwatch stopwatch = Stopwatch()..start();
-
-    for (final MapEntry<String, ParsedFileData> entry in _filesRegistry.entries) {
-      final String targetFilePath = entry.key;
-      await _generateCode(
-        targetFilePath: targetFilePath,
-        outputFilePath: targetFilePath.replaceFirst('.dart', '.gen.dart'),
-        compilationUnit: entry.value.compilationUnit,
-      );
-    }
-
-    stopwatch.stop();
-    _logger.info('~ Completed build in ${stopwatch.elapsed.inMilliseconds}ms..');
-  }
-
-  void _onWatchEvent(WatchEvent event) async {
-    final String targetFilePath = path.normalize(event.path);
-    Completer<void>? completer;
-
+class DataClassPluginGenerator extends TachyonPluginCodeGenerator {
+  DataClassPluginOptions _getPluginOptionWithDefaultFallback(String projectDirPath) {
     try {
-      if (!_dartFileNameMatcher.hasMatch(path.basename(targetFilePath))) {
-        return;
-      }
-
-      final String outputFilePath = targetFilePath.replaceFirst('.dart', '.gen.dart');
-      await _activeWrites[outputFilePath]?.future;
-
-      completer = Completer<void>();
-      _activeWrites[outputFilePath] = completer;
-
-      if (event.type == ChangeType.REMOVE) {
-        try {
-          await File(outputFilePath).delete();
-        } catch (_) {}
-
-        return;
-      }
-
-      await _generateCode(targetFilePath: targetFilePath, outputFilePath: outputFilePath);
-    } catch (error, stackTrace) {
-      _logger.exception(error, stackTrace);
-    } finally {
-      completer?.safeComplete();
+      return DataClassPluginOptions.fromFile(getDataClassPluginOptionsFile(
+        projectDirPath,
+      ));
+    } catch (_) {
+      return const DataClassPluginOptions();
     }
   }
 
-  Future<void> _generateCode({
-    required String targetFilePath,
-    required String outputFilePath,
-    bool skipDependencies = false,
-    CompilationUnit? compilationUnit,
-    String indent = '',
-    bool reportTime = true,
-  }) async {
-    final String relativeFilePath = path.relative(targetFilePath, from: directory.path);
-    final DataClassPluginOptions pluginOptions = await DataClassPluginOptions.fromFile(
-        common_utils.getDataClassPluginOptionsFile(directory.path));
-
-    if (!pluginOptions.allowedFilesGenerationPaths
-        .any((Glob glob) => glob.matches(relativeFilePath))) {
-      return;
-    }
-
-    _logger.debug('$indent~ Checking $relativeFilePath');
-
-    late final Stopwatch? stopwatch;
-    if (reportTime) {
-      stopwatch = Stopwatch()..start();
-    } else {
-      stopwatch = null;
-    }
-
-    CodeWriter codeWriter = CodeWriter.stringBuffer();
-
-    if (compilationUnit == null) {
-      _filesRegistry[targetFilePath] = ParsedFileData(
-        absolutePath: targetFilePath,
-        compilationUnit: targetFilePath.parse(featureSet: FeatureSet.latestLanguageVersion()).unit,
-        lastModifiedAt: await File(targetFilePath).lastModified(),
-      );
-    }
-
-    compilationUnit ??= _filesRegistry[targetFilePath]!.compilationUnit;
-
+  @override
+  FutureOr<String> generate(
+    FileChangeBuildInfo buildInfo,
+    TachyonDeclarationFinder declarationFinder,
+    Logger logger,
+  ) async {
+    final CodeWriter codeWriter = CodeWriter.stringBuffer();
     final ClassCollectorAstVisitor classCollectorVisitor =
         ClassCollectorAstVisitor(matcher: (ClassDeclaration node) {
       return node.metadata
           .any((Annotation meta) => meta.isDataClassAnnotation || meta.isUnionAnnotation);
     });
-    compilationUnit.visitChildren(classCollectorVisitor);
+    buildInfo.compilationUnit.visitChildren(classCollectorVisitor);
 
-    final List<ClassDeclaration> classDeclarations = classCollectorVisitor.matchedNodes;
-
-    if (classDeclarations.isEmpty) {
-      final File outputFile = File(outputFilePath);
-      if (await outputFile.exists()) {
-        try {
-          await outputFile.delete();
-        } catch (_) {}
-      }
-
-      if (!skipDependencies) {
-        await _rebuildDependants(targetFilePath: targetFilePath, indent: indent);
-      }
-
-      if (reportTime) {
-        stopwatch!.stop();
-        _logger.info(
-            '$indent~ Finished building $relativeFilePath in ${stopwatch.elapsed.inMilliseconds}ms');
-      }
-      return;
-    }
-
-    final List<ClassDeclaration> dataClasses = classDeclarations
+    final List<ClassDeclaration> dataClasses = classCollectorVisitor.matchedNodes
         .where((ClassDeclaration classDecl) => classDecl.hasDataClassAnnotation)
         .toList(growable: false);
-    final List<ClassDeclaration> unionClasses = classDeclarations
+    final List<ClassDeclaration> unionClasses = classCollectorVisitor.matchedNodes
         .where((ClassDeclaration classDecl) => classDecl.hasUnionAnnotation)
         .toList(growable: false);
 
-    _logger.debug('$indent~ Starting build for $relativeFilePath');
+    final DataClassPluginOptions pluginOptions =
+        _getPluginOptionWithDefaultFallback(buildInfo.projectDirectoryPath);
 
     final JsonKeyNameConventionGetter jsonKeyNameConventionGetter = utils.getJsonKeyNameConvention(
-      targetFileRelativePath: path.relative(targetFilePath),
+      targetFileRelativePath: path.relative(buildInfo.targetFilePath),
       pluginOptions: pluginOptions,
     );
 
-    codeWriter
-      ..writeln('// AUTO GENERATED - DO NOT MODIFY')
-      ..writeln()
-      ..writeln(
-          '// ignore_for_file: library_private_types_in_public_api, unused_element, unused_field')
-      ..writeln()
-      ..writeln("part of '${path.basename(targetFilePath)}';")
-      ..writeln();
-
     await _generateDataClasses(
+      buildInfo: buildInfo,
+      declarationFinder: declarationFinder,
       codeWriter: codeWriter,
       pluginOptions: pluginOptions,
-      compilationUnit: compilationUnit,
+      compilationUnit: buildInfo.compilationUnit,
       classDeclarations: dataClasses,
-      targetFilePath: targetFilePath,
-      targetFileRelativePath: relativeFilePath,
+      targetFilePath: buildInfo.targetFilePath,
+      targetFileRelativePath:
+          path.relative(buildInfo.targetFilePath, from: buildInfo.projectDirectoryPath),
       jsonKeyNameConventionGetter: jsonKeyNameConventionGetter,
+      logger: logger,
     );
 
     await _generateUnionClasses(
+      buildInfo: buildInfo,
+      declarationFinder: declarationFinder,
       codeWriter: codeWriter,
       pluginOptions: pluginOptions,
-      compilationUnit: compilationUnit,
+      compilationUnit: buildInfo.compilationUnit,
       classDeclarations: unionClasses,
-      targetFilePath: targetFilePath,
-      targetFileRelativePath: relativeFilePath,
+      targetFilePath: buildInfo.targetFilePath,
+      targetFileRelativePath:
+          path.relative(buildInfo.targetFilePath, from: buildInfo.projectDirectoryPath),
       jsonKeyNameConventionGetter: jsonKeyNameConventionGetter,
+      logger: logger,
     );
 
-    final String content = codeWriter.content;
-    if (content.isEmpty) {
-      try {
-        await File(outputFilePath).delete();
-      } catch (_) {}
-    } else {
-      try {
-        await File(outputFilePath).writeAsString(DartFormatter(
-          pageWidth: pluginOptions.generatedFileLineLength,
-        ).format(content));
-      } on FormatterException catch (e) {
-        _logger
-          ..error('Invalid code generation for $relativeFilePath')
-          ..writeln(e);
-      }
-    }
-
-    if (!skipDependencies) {
-      await _rebuildDependants(targetFilePath: targetFilePath, indent: indent);
-    }
-
-    if (reportTime) {
-      stopwatch!.stop();
-      _logger.info(
-          '$indent~ Finished building $relativeFilePath in ${stopwatch.elapsed.inMilliseconds}ms');
-    }
-  }
-
-  Future<void> _rebuildDependants({
-    required final String targetFilePath,
-    required final String indent,
-  }) async {
-    final List<String> dependants = _dependencyGraph.getDependants(targetFilePath);
-    if (dependants.isEmpty) {
-      return;
-    }
-
-    _logger.debug('  Rebuilding ${dependants.length} depandants..');
-    await Future.wait(<Future<void>>[
-      for (final String dependency in dependants)
-        _generateCode(
-          targetFilePath: dependency,
-          outputFilePath: dependency.replaceFirst('.dart', '.gen.dart'),
-          compilationUnit: _filesRegistry[dependency]?.compilationUnit,
-          indent: '  $indent',
-          skipDependencies: true,
-          reportTime: false,
-        )
-    ]);
+    return codeWriter.content;
   }
 
   Future<void> _generateDataClasses({
-    required CodeWriter codeWriter,
-    required DataClassPluginOptions pluginOptions,
-    required CompilationUnit compilationUnit,
-    required List<ClassDeclaration> classDeclarations,
-    required String targetFilePath,
-    required String targetFileRelativePath,
-    required JsonKeyNameConventionGetter jsonKeyNameConventionGetter,
+    required final FileChangeBuildInfo buildInfo,
+    required final TachyonDeclarationFinder declarationFinder,
+    required final CodeWriter codeWriter,
+    required final DataClassPluginOptions pluginOptions,
+    required final CompilationUnit compilationUnit,
+    required final List<ClassDeclaration> classDeclarations,
+    required final String targetFilePath,
+    required final String targetFileRelativePath,
+    required final JsonKeyNameConventionGetter jsonKeyNameConventionGetter,
+    required final Logger logger,
   }) async {
     for (final ClassDeclaration classDeclaration in classDeclarations) {
       final String className = classDeclaration.name.lexeme;
@@ -476,14 +170,8 @@ class CodeGenerator {
           generatedClassName: generatedClassName,
           classTypeParametersWithoutConstraints: classTypeParametersWithoutConstraints,
           jsonKeyNameConventionGetter: jsonKeyNameConventionGetter,
-          classDeclarationFinder: (String name) {
-            return _declarationFinder.findClassOrEnumDeclarationByName(
-              name,
-              compilationUnit: compilationUnit,
-              targetFilePath: targetFilePath,
-            );
-          },
-          logger: _logger,
+          classDeclarationFinder: declarationFinder.findClassOrEnum,
+          logger: logger,
         ).execute();
       }
 
@@ -493,14 +181,8 @@ class CodeGenerator {
           codeWriter: codeWriter,
           fields: fields,
           jsonKeyNameConventionGetter: jsonKeyNameConventionGetter,
-          classDeclarationFinder: (String name) {
-            return _declarationFinder.findClassOrEnumDeclarationByName(
-              name,
-              compilationUnit: compilationUnit,
-              targetFilePath: targetFilePath,
-            );
-          },
-          logger: _logger,
+          classDeclarationFinder: declarationFinder.findClassOrEnum,
+          logger: logger,
         ).execute();
       }
 
@@ -541,15 +223,11 @@ class CodeGenerator {
       // If that case is true then override the config for copyWith for the current class
       // as it would not make sense to have copyWith for super class and not a subclass
       if (classDeclaration.extendsClause?.superclass != null) {
-        final String superClassName = classDeclaration.extendsClause!.superclass.name.name;
+        final String superClassName = classDeclaration.extendsClause!.superclass.name2.lexeme;
         final ClassOrEnumDeclarationMatch? superClassMatch =
-            await _declarationFinder.findClassOrEnumDeclarationByName(
-          superClassName,
-          compilationUnit: compilationUnit,
-          targetFilePath: targetFilePath,
-        );
-
+            await declarationFinder.findClassOrEnum(superClassName);
         final NamedCompilationUnitMember? superClassDeclaration = superClassMatch?.node;
+
         if (superClassDeclaration is ClassDeclaration &&
             superClassDeclaration.hasDataClassAnnotation) {
           final AnnotationValueExtractor annotationValueExtractor = AnnotationValueExtractor(
@@ -562,13 +240,12 @@ class CodeGenerator {
         }
 
         if (supportsCopyWith != null) {
-          _logger.warning(
+          logger.warning(
               '~ Overriden copyWith configuration for class <$className> based on super class <$superClassName>');
         }
       }
       supportsCopyWith ??= dataClassAnnotationValueExtractor.getBool('copyWith') ??
-          pluginOptions.dataClass
-              .effectiveCopyWith(path.relative(targetFileRelativePath, from: directory.path));
+          pluginOptions.dataClass.effectiveCopyWith(targetFileRelativePath);
 
       if (supportsCopyWith) {
         await CopyWithGenerator(
@@ -578,14 +255,8 @@ class CodeGenerator {
           generatedClassName: generatedClassName,
           classTypeParametersSource: classTypeParametersSource,
           classTypeParametersWithoutConstraints: classTypeParametersWithoutConstraints,
-          classDeclarationFinder: (String name) {
-            return _declarationFinder.findClassOrEnumDeclarationByName(
-              name,
-              compilationUnit: compilationUnit,
-              targetFilePath: targetFilePath,
-            );
-          },
-          projectDirectoryPath: directory.path,
+          classDeclarationFinder: declarationFinder.findClassOrEnum,
+          projectDirectoryPath: buildInfo.projectDirectoryPath,
           pluginOptions: pluginOptions,
         ).execute();
 
@@ -600,13 +271,16 @@ class CodeGenerator {
   }
 
   Future<void> _generateUnionClasses({
-    required CodeWriter codeWriter,
-    required DataClassPluginOptions pluginOptions,
-    required CompilationUnit compilationUnit,
-    required List<ClassDeclaration> classDeclarations,
-    required String targetFilePath,
-    required String targetFileRelativePath,
-    required JsonKeyNameConventionGetter jsonKeyNameConventionGetter,
+    required final FileChangeBuildInfo buildInfo,
+    required final TachyonDeclarationFinder declarationFinder,
+    required final CodeWriter codeWriter,
+    required final DataClassPluginOptions pluginOptions,
+    required final CompilationUnit compilationUnit,
+    required final List<ClassDeclaration> classDeclarations,
+    required final String targetFilePath,
+    required final String targetFileRelativePath,
+    required final JsonKeyNameConventionGetter jsonKeyNameConventionGetter,
+    required final Logger logger,
   }) async {
     for (final ClassDeclaration classDeclaration in classDeclarations) {
       final String className = classDeclaration.name.lexeme;
@@ -634,13 +308,16 @@ class CodeGenerator {
       final AnnotationValueExtractor unionAnnotationValueExtractor =
           AnnotationValueExtractor(classDeclaration.unionAnnotation);
 
-      UnionWhenGenerator(
-        codeWriter: codeWriter,
-        className: className,
-        classTypeParametersSource: classTypeParametersSource,
-        classTypeParametersWithoutConstraints: classTypeParametersWithoutConstraints,
-        factoriesWithRedirectedConstructors: factoriesWithRedirectedConstructors,
-      ).execute();
+      if (unionAnnotationValueExtractor.getBool('when') ??
+          pluginOptions.union.effectiveWhen(targetFileRelativePath)) {
+        UnionWhenGenerator(
+          codeWriter: codeWriter,
+          className: className,
+          classTypeParametersSource: classTypeParametersSource,
+          classTypeParametersWithoutConstraints: classTypeParametersWithoutConstraints,
+          factoriesWithRedirectedConstructors: factoriesWithRedirectedConstructors,
+        ).execute();
+      }
 
       if (unionAnnotationValueExtractor.getBool('fromJson') ??
           pluginOptions.union.effectiveFromJson(targetFileRelativePath)) {
@@ -705,31 +382,24 @@ class CodeGenerator {
             generatedClassName: generatedClassName,
             classTypeParametersWithoutConstraints: classTypeParametersWithoutConstraints,
             jsonKeyNameConventionGetter: jsonKeyNameConventionGetter,
-            classDeclarationFinder: (String name) {
-              return _declarationFinder.findClassOrEnumDeclarationByName(
-                name,
-                compilationUnit: compilationUnit,
-                targetFilePath: targetFilePath,
-              );
-            },
-            logger: _logger,
+            classDeclarationFinder: declarationFinder.findClassOrEnum,
+            logger: logger,
           ).execute();
         }
 
         if (unionAnnotationValueExtractor.getBool('toJson') ??
             pluginOptions.union.effectiveToJson(targetFileRelativePath)) {
+          final String? toJsonNamedKeyToFactoryName =
+              unionAnnotationValueExtractor.getString('toJsonNamedKeyToFactoryName');
           await ToJsonGenerator(
             codeWriter: codeWriter,
             fields: fields,
             jsonKeyNameConventionGetter: jsonKeyNameConventionGetter,
-            classDeclarationFinder: (String name) {
-              return _declarationFinder.findClassOrEnumDeclarationByName(
-                name,
-                compilationUnit: compilationUnit,
-                targetFilePath: targetFilePath,
-              );
-            },
-            logger: _logger,
+            namedKeyToFactoryEntry: toJsonNamedKeyToFactoryName != null
+                ? "'$toJsonNamedKeyToFactoryName': '${ctor.name!.lexeme}'"
+                : null,
+            classDeclarationFinder: declarationFinder.findClassOrEnum,
+            logger: logger,
           ).execute();
         }
 
@@ -761,8 +431,7 @@ class CodeGenerator {
         codeWriter.writeln('}');
 
         if (unionAnnotationValueExtractor.getBool('copyWith') ??
-            pluginOptions.union
-                .effectiveCopyWith(path.relative(targetFileRelativePath, from: directory.path))) {
+            pluginOptions.union.effectiveCopyWith(targetFileRelativePath)) {
           await CopyWithGenerator(
             codeWriter: codeWriter,
             fields: fields,
@@ -770,14 +439,8 @@ class CodeGenerator {
             generatedClassName: generatedClassName,
             classTypeParametersSource: classTypeParametersSource,
             classTypeParametersWithoutConstraints: classTypeParametersWithoutConstraints,
-            classDeclarationFinder: (String name) {
-              return _declarationFinder.findClassOrEnumDeclarationByName(
-                name,
-                compilationUnit: compilationUnit,
-                targetFilePath: targetFilePath,
-              );
-            },
-            projectDirectoryPath: directory.path,
+            classDeclarationFinder: declarationFinder.findClassOrEnum,
+            projectDirectoryPath: buildInfo.projectDirectoryPath,
             pluginOptions: pluginOptions,
           ).execute();