+++ package/lib/src/contributors/generators/copy_with.dart
@@ -1,7 +1,7 @@
 import 'package:analyzer/dart/element/element.dart';
-import 'package:data_class_plugin/src/backend/core/generators/generator.dart';
-import 'package:data_class_plugin/src/common/code_writer.dart';
+import 'package:data_class_plugin/src/common/generator.dart';
 import 'package:data_class_plugin/src/extensions/extensions.dart';
+import 'package:tachyon/tachyon.dart';
 
 class CopyWithGenerator implements Generator {
   CopyWithGenerator({