+++ package/test/contributors/annotations/data_class/test_files/in_4_to_json.dart
@@ -1,0 +1,10 @@
+import 'package:data_class_plugin/data_class_plugin.dart';
+
+@DataClass(
+  toJson: true,
+)
+class User {
+  int get id;
+
+  String get username;
+}