+++ package/test/contributors/annotations/data_class/test_files/in_5_default_value.dart
@@ -1,0 +1,11 @@
+import 'package:data_class_plugin/data_class_plugin.dart';
+
+@DataClass()
+class User {
+  int get id;
+
+  String get username;
+
+  @DefaultValue('')
+  String get email;
+}