+++ package/test/contributors/annotations/union/test_files/in_1_simple.dart
@@ -1,0 +1,7 @@
+import 'package:data_class_plugin/data_class_plugin.dart';
+
+@Union()
+class AsyncResult {
+  factory AsyncResult.data() = AsyncResultData;
+  factory AsyncResult.error() = AsyncResultError;
+}