+++ package/test/contributors/annotations/union/test_files/in_5_to_json.dart
@@ -1,0 +1,9 @@
+import 'package:data_class_plugin/data_class_plugin.dart';
+
+@Union(
+  toJson: true,
+)
+class AsyncResult {
+  factory AsyncResult.data() = AsyncResultData;
+  factory AsyncResult.error() = AsyncResultError;
+}