+++ package/test/contributors/annotations/union/test_files/out_1_simple.dart
@@ -1,0 +1,9 @@
+part 'in_1_simple.gen.dart';
+
+@Union()
+abstract class AsyncResult {
+  const AsyncResult._();
+
+  factory AsyncResult.data() = AsyncResultData;
+  factory AsyncResult.error() = AsyncResultError;
+}