+++ package/test/contributors/annotations/union/test_files/out_4_from_json.dart
@@ -1,0 +1,14 @@
+part 'in_4_from_json.gen.dart';
+
+@Union(
+  fromJson: true,
+)
+abstract class AsyncResult {
+  const AsyncResult._();
+
+  /// Creates an instance of [AsyncResult] from [json]
+  factory AsyncResult.fromJson(Map<dynamic, dynamic> json) => _$AsyncResultFromJson(json);
+
+  factory AsyncResult.data() = AsyncResultData;
+  factory AsyncResult.error() = AsyncResultError;
+}