+++ package/tools/analyzer_plugin/bin/plugin.dart
@@ -1,6 +1,6 @@
 import 'dart:isolate';
 
-import 'package:data_class_plugin/plugin.dart';
+import 'package:data_class_plugin/analyzer_plugin.dart';
 
 void main(List<String> args, SendPort sendPort) {
   start(args, sendPort);